Fixed session stuff, /label now protected
